+++ Raether/src/Box.h
@@ -5,19 +5,18 @@
 
 class Box : public Hittable {
 public:
-	Box(const glm::vec3& a, const glm::vec3& b, std::shared_ptr<Material> matid);
-	Box(const glm::vec3& origin, std::shared_ptr<Material> matid, const glm::vec3& boxwidth);
+	Box(const glm::dvec3& a, const glm::dvec3& b, std::shared_ptr<Material> matid);
+	Box(const glm::dvec3& origin, std::shared_ptr<Material> matid, const glm::dvec3& boxwidth);
 
 	bool Hit(const Ray& ray, Interval hitdist, Hitrec& hitrecord) const override;
 
 	Aabb BoundingBox() const override { return bbox; }
 
 private:
-	glm::vec3 Min;
-	glm::vec3 Max;
+	glm::dvec3 Min;
+	glm::dvec3 Max;
 	std::shared_ptr<Material> MaterialId;
 	Aabb bbox;
 	std::vector<std::shared_ptr<Hittable>> Quads;
 	std::shared_ptr<BvhNode> RootNode;
 };
-
