+++ Raether/src/Texture.h
@@ -10,19 +10,19 @@
 class Texture {
 public:
 	virtual ~Texture() = default;
-	virtual glm::vec3 value(double u, double v, glm::vec3& point) const = 0;
+	virtual glm::dvec3 value(double u, double v, glm::dvec3& point) const = 0;
 };
 
 class SolidColor : public Texture {
 public:
-	SolidColor(glm::vec3& color) : Albedo(color) {}
-	SolidColor(float r, float g, float b) : SolidColor(glm::vec3(r, g, b)) {}
+	SolidColor(glm::dvec3& color) : Albedo(color) {}
+	SolidColor(double r, double g, double b) : SolidColor(glm::dvec3(r, g, b)) {}
 
-	glm::vec3 value(double u, double v, glm::vec3& point) const override {
+	glm::dvec3 value(double u, double v, glm::dvec3& point) const override {
 		return Albedo;
 	}
 private:
-	glm::vec3 Albedo;
+	glm::dvec3 Albedo;
 };
 
 class CheckerTexture : public Texture {
@@ -32,11 +32,11 @@ public:
 		Even(even),
 		Odd(odd)
 	{}
-	CheckerTexture(double scale, glm::vec3& color1, glm::vec3& color2) :
+	CheckerTexture(double scale, glm::dvec3& color1, glm::dvec3& color2) :
 		CheckerTexture(scale, std::make_shared<SolidColor>(color1), std::make_shared<SolidColor>(color2))
 	{}
 
-	glm::vec3 value(double u, double v, glm::vec3& point) const override {
+	glm::dvec3 value(double u, double v, glm::dvec3& point) const override {
 		int xInteger = int(glm::floor(InvScale * point.x));
 		int yInteger = int(glm::floor(InvScale * point.y));
 		int zInteger = int(glm::floor(InvScale * point.z));
@@ -57,21 +57,21 @@ public:
 	ImageTexture() : Img("UVChecker_2K.png") {}
 	ImageTexture(const char* imagePath) : Img(imagePath) {}
 
-	glm::vec3 value(double u, double v, glm::vec3& point) const override {
+	glm::dvec3 value(double u, double v, glm::dvec3& point) const override {
 
 		// If no image is loaded return color-cyan
-		if (Img.GetWidth() == 0) return glm::vec3(0.f, 1.f, 1.f);
-		
+		if (Img.GetWidth() == 0) return glm::dvec3(0.0, 1.0, 1.0);
+
 		u = Interval(0.0, 1.0).Clamp(u);
 		v = 1.0 - Interval(0.0, 1.0).Clamp(v);
 
 		auto i = int(u * Img.GetWidth());
 		auto j = int(v * Img.GetHeight());
 
-		float colorScale = 1.f / 255.f;
+		double colorScale = 1.0 / 255.0;
 		const unsigned char* pixel = Img.PixelData(i, j);
 
-		return glm::vec3(colorScale * pixel[0], colorScale * pixel[1], colorScale * pixel[2]);
+		return glm::dvec3(colorScale * pixel[0], colorScale * pixel[1], colorScale * pixel[2]);
 	}
 
 private:
@@ -80,9 +80,9 @@ private:
 
 class NoiseTexture : public Texture {
 public:
-	NoiseTexture(double scale) : InvScale( 1.0 / scale){}
+	NoiseTexture(double scale) : InvScale(1.0 / scale) {}
 
-	glm::vec3 value(double u, double v, glm::vec3& point) const override {
+	glm::dvec3 value(double u, double v, glm::dvec3& point) const override {
 		return glm::dvec3(0.5) * (1.0 + glm::sin(InvScale * point.z + 10 * Noise.Turbulence(glm::dvec3(point), 7)));
 	}
